python-service/requirements.txt
# Fast YOLO detection service
fastapi==0.115.0
uvicorn[standard]==0.32.0
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
pillow==11.0.0
numpy==1.26.4
pydantic==2.9.2
